+++ frontend/src/components/navigation/index.ts
@@ -1,2 +1,11 @@
 export { default as AppNavigation } from './AppNavigation'
 export type { AppNavigationProps } from './AppNavigation'
+
+export { default as AppDrawerContent } from './AppDrawerContent'
+export type { AppDrawerContentProps } from './AppDrawerContent'
+
+export { default as AppDrawerHeader } from './AppDrawerHeader'
+export type { AppDrawerHeaderProps } from './AppDrawerHeader'
+
+export { default as NavigationMenu } from './NavigationMenu'
+export type { NavigationMenuProps } from './NavigationMenu'
